ver 2

+++ profile.php
@@ -1,25 +1,70 @@
-<!-- <?php
-include('./classes/DB.php');
-include('./classes/Login.php');
-include('./classes/Post.php');
-include('./classes/Comment.php');
+<?php
+include ('classes/DB.php');
+include('classes/Login.php');
+include ('classes/Post.php');
+include('./classes/Image.php');
 
-$showTimeline = False;
-  if(Login::isLoggedIn()){
-    $userid = Login::isLoggedIn();
-    $showTimeline = True;
-  }
-  else {
-    die("not Logged in");
-  }
-  if (isset($_GET['postid'])) {
-          Post::likePost($_GET['postid'], $userid);
-  }
-  if (isset($_POST['comment'])){
-    Comment::createComment($_POST['commentbody'],$_GET['postid'],$userid);
-  }
+$username = '';
+$isFollowing = False;
+$isVerified = False;
 
-  if (isset($_POST['deletepost'])) {
+if(isset($_GET['username'])){
+    if(DB::query('SELECT username FROM users WHERE username=:username',array(':username'=>$_GET['username']))) {
+
+        $username = DB::query('SELECT username FROM users WHERE username=:username',array(':username'=>$_GET['username']))[0]['username'];
+
+        $userid = DB::query('SELECT id from users where username=:username',array(':username'=>$_GET['username']))[0]['id'];
+
+        $followerid = Login::isLoggedIn();
+
+        $isVerified = DB::query('SELECT verified FROM users WHERE username=:username',array(':username'=>$_GET['username']))[0]['verified'];
+
+        if(isset($_POST['follow'])){
+
+          if ($userid != $followerid) {
+
+            if(!DB::query('SELECT follower_id FROM followers WHERE user_id = :userid AND follower_id = :followerid',array(':userid'=>$userid,':followerid'=>$followerid))) {
+
+                DB::query('INSERT INTO followers (user_id,follower_id) VALUES(:userid,:followerid)',array(':userid'=>$userid,':followerid'=>$followerid));
+            }
+
+            else{
+
+              echo 'Already Following';
+
+            }
+
+            $isFollowing = True;
+
+          }
+
+        }
+
+        if(isset($_POST['unfollow'])){
+
+          if ($userid != $followerid) {
+
+            if(DB::query('SELECT follower_id FROM followers WHERE user_id = :userid AND follower_id = :followerid',array(':userid'=>$userid,':followerid'=>$followerid))){
+
+                DB::query('DELETE FROM followers WHERE user_id = :userid AND follower_id = :followerid',array(':userid'=>$userid,':followerid'=>$followerid));
+
+            }
+
+            $isFollowing = False;
+
+        }
+
+      }
+
+
+        if (DB::query('SELECT follower_id FROM followers WHERE user_id=:userid AND follower_id=:followerid', array(':userid'=>$userid, ':followerid'=>$followerid))) {
+
+                      //echo 'Already following!';
+                      $isFollowing = True;
+
+      }
+
+      if (isset($_POST['deletepost'])) {
                       if (DB::query('SELECT id FROM posts WHERE id=:postid AND user_id=:userid', array(':postid'=>$_GET['postid'], ':userid'=>$followerid))) {
                               DB::query('DELETE FROM posts WHERE id=:postid and user_id=:userid', array(':postid'=>$_GET['postid'], ':userid'=>$followerid));
                               DB::query('DELETE FROM post_likes WHERE post_id=:postid', array(':postid'=>$_GET['postid']));
@@ -27,42 +72,70 @@ $showTimeline = False;
                       }
               }
 
-  $followingposts = DB::query('SELECT posts.id, posts.body, posts.likes, users.`username` FROM users, posts, followers
-  WHERE posts.user_id = followers.user_id
-  AND users.id = posts.user_id
-  AND follower_id = :userid
-  ORDER BY posts.likes DESC;',array(':userid'=>$userid));
 
+      if (isset($_POST['post'])) {
 
-  foreach($followingposts as $post) {
-        echo $post['body']." ~ ".$post['username'];
-        echo "<form action='index.php?postid=".$post['id']."' method='post'>";
-        if (!DB::query('SELECT post_id FROM post_likes WHERE post_id=:postid AND user_id=:userid', array(':postid'=>$post['id'], ':userid'=>$userid))) {
-        echo "<input type='submit' name='like' value='Like'>";
-        } else {
-        echo "<input type='submit' name='unlike' value='Unlike'>";
+            if ($_FILES['postimg']['size'] == 0) {
+                                Post::createPost($_POST['postbody'], Login::isLoggedIn(), $userid);
+                                header('Location: '.$_SERVER['REQUEST_URI']);
+              } else {
+                            $postid = Post::createImgPost($_POST['postbody'], Login::isLoggedIn(), $userid);
+                            Image::uploadImage('postimg', "UPDATE posts SET postimg=:postimg WHERE id=:postid", array(':postid'=>$postid));
+                            header('Location: '.$_SERVER['REQUEST_URI']);
+
+                }
+
         }
-        echo "<span>".$post['likes']." likes</span>
-        </form>
-        <form action='index.php?postid=".$post['id']."' method='post'>
-        <textarea name='commentbody' rows='3' cols='50'></textarea>
-        <input type='submit' name='comment' value='Comment'>
-        </form>
-        ";
-        Comment::displayComments($post['id']);
-        echo "<hr /></br />";
-}
 
-?> -->
+        if(isset($_GET['postid']) && !isset($_POST['deletepost'])){
+          Post::likePost($_GET['postid'],$followerid);
+        }
 
 
+        $posts = Post::getPosts($userid,$username,$followerid);
+
+      }
+
+    else {
+
+      die("user doesn't exist");
+
+      exit();
+    }
+
+  }
+
+  //TODO Add a function to change verified to 1
+  //TODO Follow button disappears after posting
+
+?>
+
+<!-- <h1><?php echo $username; ?></h1>
+<h4><?php if($isVerified) {echo '- verified user'; } ?></h4>
+<form action="profile.php?username=<?php echo $username; ?>" method="post"> -->
 <?php
-  if(Login::isLoggedIn()){
-   $userid = Login::isLoggedIn();
-   $username=DB::query('SELECT username FROM users WHERE id=:userid',array(":userid"=>$userid))[0]['username'];
-   $proflink;
+    if ($userid != $followerid) {
+    if(!$isFollowing) {
+      echo '<input type="submit" name="follow" value="Follow">';
+    } else {
+      echo '<input type="submit" name="unfollow" value="Unfollow">';
+    }
   }
 ?>
+</form>
+<!--
+<form action="profile.php?username=<?php echo $username; ?>" method="post" enctype="multipart/form-data">
+        <textarea name="postbody" rows="8" cols="80"></textarea>
+        <br />Upload an image:
+        <input type="file" name="postimg">
+        <input type="submit" name="post" value="Post">
+</form>
+
+<div class="posts">
+  <?php echo $posts; ?>
+
+</div>
+ -->
 
 
 
@@ -77,6 +150,7 @@ $showTimeline = False;
     <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
     <link rel="stylesheet" href="assets/fonts/ionicons.min.css">
     <link rel="stylesheet" href="assets/css/Footer-Dark.css">
+    <link rel="stylesheet" href="assets/css/Highlight-Clean.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/aos/2.1.1/aos.css">
     <link rel="stylesheet" href="assets/css/Login-Form-Clean.css">
     <link rel="stylesheet" href="assets/css/Navigation-Clean1.css">
@@ -85,8 +159,7 @@ $showTimeline = False;
 </head>
 
 <body>
-  <header class="hidden-sm hidden-md hidden-lg">
-
+    <header class="hidden-sm hidden-md hidden-lg">
         <div class="searchbox">
             <form>
                 <h1 class="text-left">Pixter</h1>
@@ -98,10 +171,9 @@ $showTimeline = False;
                     <ul class="dropdown-menu dropdown-menu-right" role="menu">
                         <li role="presentation"><a href="#">My Profile</a></li>
                         <li class="divider" role="presentation"></li>
-                        <li role="presentation"><a href="#">Timeline </a></li>
-                        <li role="presentation"><a href="profile.php?username=<?php echo $username; ?>">Profile </a></li>
+                        <li role="presentation"><a href="index.php">Timeline </a></li>
                         <li role="presentation"><a href="#">My Account</a></li>
-                        <li role="presentation" ><a href="#" class="logout">Logout </a></li>
+                        <li role="presentation"><a href="#" class="logout">Logout </a></li>
                     </ul>
                 </div>
             </form>
@@ -121,27 +193,26 @@ $showTimeline = False;
                         </div>
                     </form>
                     <ul class="nav navbar-nav hidden-md hidden-lg navbar-right">
-                        <li role="presentation"><a href="#">My Timeline</a></li>
+                        <li role="presentation"><a href="index.php">My Timeline</a></li>
                         <li class="dropdown open"><a class="dropdown-toggle" data-toggle="dropdown" aria-expanded="true" href="#"><?php echo $username; ?> <span class="caret"></span></a>
                             <ul class="dropdown-menu dropdown-menu-right" role="menu">
-                                <li role="presentation"><a href="profile.php?username=<?php echo $username; ?>">My Profile</a></li>
+                                <li role="presentation"><a href="#">My Profile</a></li>
                                 <li class="divider" role="presentation"></li>
-                                <li role="presentation"><a href="#">Timeline </a></li>
+                                <li role="presentation"><a href="index.php">Timeline </a></li>
                                 <li role="presentation"><a href="#">My Account</a></li>
                                 <li role="presentation"><a href="#" class="logout">Logout </a></li>
                             </ul>
                         </li>
                     </ul>
                     <ul class="nav navbar-nav hidden-xs hidden-sm navbar-right">
-                        <li class="active" role="presentation"><a href="#">Timeline</a></li>
-                        <li role="presentation"><a href="profile.php?username=<?php echo $username; ?>">Profile</a></li>
+                        <li role="presentation"><a href="index.php">Timeline</a></li>
+                        <li class="active" role="presentation"><a href="#">Profile</a></li>
                         <li class="dropdown"><a class="dropdown-toggle" data-toggle="dropdown" aria-expanded="false" href="#"><?php echo $username; ?> <span class="caret"></span></a>
                             <ul class="dropdown-menu dropdown-menu-right" role="menu">
-                                <li role="presentation"><a href="profile.php?username=<?php echo $username; ?>">My Profile</a></li>
+                                <li role="presentation"><a href="#">My Profile</a></li>
                                 <li class="divider" role="presentation"></li>
-                                <li role="presentation"><a href="#">Timeline </a></li>
-                                <li role="presentation"><a href="<?php echo"profile.php?username=$username"; ?> ">Profile</a></li>
-                                <li role="presentation"><a href="login.html">My Account</a></li>
+                                <li role="presentation"><a href="index.php">Timeline </a></li>
+                                <li role="presentation"><a href="#">My Account</a></li>
                                 <li role="presentation"><a href="#" class="logout">Logout </a></li>
                             </ul>
                         </li>
@@ -151,8 +222,29 @@ $showTimeline = False;
         </nav>
     </div>
     <div class="container">
-        <h1>Timeline </h1>
-        <div class="timelineposts">
+        <h1><?php echo $username; ?><?php if($isVerified){echo '<i class="glyphicon glyphicon-ok-sign verified" data-toggle="tooltip" title="Verified User" style="font-size:28px;color:#da052b;"></i>';} ?></h1></div>
+    <div>
+        <div class="container">
+            <div class="row">
+                <div class="col-md-3">
+                    <ul class="list-group">
+                        <li class="list-group-item"><span><strong>About Me</strong></span>
+                            <p>Welcome to <?php echo $username; ?>'s profile</p>
+                        </li>
+                    </ul>
+                </div>
+                <div class="col-md-6">
+                    <ul class="list-group">
+                      <div class="timelineposts">
+
+                      </div>
+                    </ul>
+                </div>
+                <div class="col-md-3">
+                    <button class="btn btn-default" id="newPost" type="button" style="width:100%;background-image:url(&quot;none&quot;);background-color:#da052b;color:#fff;padding:16px 32px;margin:0px 0px 6px;border:none;box-shadow:none;text-shadow:none;opacity:0.9;text-transform:uppercase;font-weight:bold;font-size:13px;letter-spacing:0.4px;line-height:1;outline:none;" >NEW POST</button>
+                    <ul class="list-group"></ul>
+                </div>
+            </div>
         </div>
     </div>
     <div class="modal fade" role="dialog" tabindex="-1" style="padding-top:100px;">
@@ -170,24 +262,6 @@ $showTimeline = False;
             </div>
         </div>
     </div>
-
-    <!-- logout-modal -->
-    <!-- <div class="modal fade" role="dialog" tabindex="-1" style="padding-top:100px;">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">×</span></button>
-                    <h4 class="modal-title">Logout</h4></div>
-                <div class="modal-body" style="max-height: 400px; overflow-y: auto">
-                    <p>The content of your modal.</p>
-                </div>
-                <div class="modal-footer">
-                    <button class="btn btn-default" type="button" data-dismiss="modal">Close</button>
-                </div>
-            </div>
-        </div>
-    </div> -->
-
     <div class="footer-dark" style="postion: relative">
         <footer>
             <div class="container">
@@ -202,11 +276,10 @@ $showTimeline = False;
     <script type="text/javascript">
 
         $(document).ready(function() {
-
                 $.ajax({
 
                         type: "GET",
-                        url: "api/posts",
+                        url: "api/profileposts?username=<?php echo $username; ?>",
                         processData: false,
                         contentType: "application/json",
                         data: '',
@@ -227,7 +300,6 @@ $showTimeline = False;
                                               '<li class="list-group-item"><blockquote><p>'+posts[index].PostBody+'</p><img src="" data-imagesrc="'+posts[index].PostImage+'" class="postimg" id="img'+posts[index].PostId+'"><footer>Posted by '+posts[index].PostedBy+' on '+posts[index].PostDate+'<button class="btn btn-default" type="button" style="color:#eb3b60;background-image:url(&quot;none&quot;);background-color:transparent;" data-id=\"'+posts[index].PostId+'\"> <i class="glyphicon glyphicon-heart" data-aos="flip-right"></i><span> '+posts[index].Likes+' Likes</span></button><button class="btn btn-default comment" data-postid=\"'+posts[index].PostId+'\" type="button" style="color:#eb3b60;background-image:url(&quot;none&quot;);background-color:transparent;"><i class="glyphicon glyphicon-flash" style="color:#f9d616;"></i><span style="color:#f9d616;"> Comments</span></button></footer></blockquote></li>'
                                             )
                                 }
-
 
                                         $('[data-postid]').click(function() {
                                                 var buttonid = $(this).attr('data-postid');
@@ -270,15 +342,17 @@ $showTimeline = False;
                                                 });
                                         })
                                 })
+
                                 $('.postimg').each(function(){
                                   this.src=$(this).attr('data-imagesrc')
                                   this.onload = function() {
                                     this.style.opacity = '1';
                                   }
+
                                 })
 
-
                         },
+
                         error: function(r) {
                                 console.log(r)
                         }
@@ -288,12 +362,13 @@ $showTimeline = False;
         });
 
 
-
         $('.logout').click(function(){
           $('.modal').modal('show')
           $('.modal-title').text(" " +'Logout')
           var res = '<form action="index.php" method="post"><input type="checkbox" name="alldevices">Sign out of All Devices</br><button type="button" class="btn btn-primary" id="logout-btn">Logout</button></form>'
+          var button = '<button class="btn btn-default" type="button" data-dismiss="modal">Close</button>'
           $('.modal-body').html(res)
+          $('.modal-footer').html(button)
           var check = 0
           $('#logout-btn').click(function(){
             if($('input[name="alldevices"]').prop("checked")==true){
@@ -320,9 +395,20 @@ $showTimeline = False;
 
         });
 
+        $('#newPost').click(function(){
+          var res = '<form action="profile.php?username=<?php echo $username; ?>" method="post" enctype="multipart/form-data"><textarea name="postbody" rows="8" cols="70"></textarea><br />Upload an image:<input type="file" name="postimg"><input type="submit" name="post" value="Post"></form>';
+          // var button = ''
+          $('.modal').modal('show')
+          $('.modal-title').text(" " +'New post')
+          $('.modal-body').html(res)
+          // $('.modal-footer').html(button)
+
+        });
+
         function showCommentsModal(res) {
                 $('.modal').modal('show')
                 $('.modal-title').text(" " +'Comments')
+                var button = '<button class="btn btn-default" type="button" data-dismiss="modal">Close</button>'
                 var output = "";
                 for (var i = 0; i < res.length; i++) {
                         output += res[i].Comment;
@@ -332,10 +418,13 @@ $showTimeline = False;
                 }
 
                 $('.modal-body').html(output)
+                $('.modal-footer').html(button)
+
         }
 
-    </script>
 
+
+    </script>
 </body>
 
 </html>
